refactor(domain): MemoRepository 프로토콜에 누락 API 추가 및 actor 제약 해제
- 프로토콜 제약을 Actor에서 Sendable로 변경 — 구현체를 actor로 강제하지 않으면서 스레드 안전 보장은 유지
- 구현체에만 있던 memoUpdatedPublisher를 프로토콜 요구사항으로 추가
- 구현체에만 있던 배열 일괄 처리 메서드 3종(moveToTrash·restore·setFavorite)을 프로토콜에 추가
- MemoUpdateType을 MemoRepositoryImpl 중첩 타입에서 Domain 최상위 타입으로 이동
