+++ Projects/Data/Sources/Repositories/MemoRepositoryImpl.swift
@@ -29,28 +29,6 @@ public enum MemoEntityError: LocalizedError {
 
 public actor MemoRepositoryImpl: MemoRepository {
     
-    public enum MemoUpdateType: Equatable {
-        public enum UpdateAttribute: Equatable {
-            case favorite(memoIDs: [UUID])
-            case titleText(memoIDs: [UUID])
-            case category(memoIDs: [UUID])
-
-            public var memoIDs: [UUID] {
-                switch self {
-                case .favorite(let memoIDs): return memoIDs
-                case .titleText(let memoIDs): return memoIDs
-                case .category(let memoIDs): return memoIDs
-                }
-            }
-        }
-        
-        case create
-        case trash
-        case delete
-        case restore
-        case update(content: UpdateAttribute)
-    }
-    
     private let context: NSManagedObjectContext
 
     public init(stack: CoreDataStack) {
